requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
cryptography==41.0.7
supabase==2.3.1
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0